Let's see if this helps
With these changes I'm aiming to suppress errors.

First of all, adding a CHECK_TICK inside the FOV was a very bad idea, I
thought it'd push the proc into a thread if the server was lagging,
instead it just made an additional call for no reason and then ran the
script, and sometimes stopped it which caused LAG

Second, I've made it so the sunlight only updates 50 lights at a time on
each time it's fired. Which will cause the lights to render much slower,
but they won't try to do them all at the same time, this means the map
will look "ugly" at times, because it's still rendering the game, but
after that, it'll look normal again.

Specifying more things into the vision cone, now the
"update_vision_cone()" will only be called with humans (We removed the
possibility of other living beings, also some attempts at reducing the
number of calls to that function. And at the same time "InCone" will
only check for living beings (I hope), which should reduce lag, since
it's less things to check.

And finally, attempts at suppressing errors.

I want to comment that there are many problems with icons, map spawns,
and stuff like that... We have to solve those ASAP with the newer maps!!